refactor: move http-transport + auth-middleware into transport/ (Plan E)
http-transport.ts is renamed to http.ts since the transport/ dir
already conveys the category.
