+++ src/server/index.ts
@@ -8,7 +8,7 @@ export {
   type InstanceRateLimit,
 } from "../resilience/adaptive-rate-limiter.js";
 export { type RateLimitConfig, RateLimiter } from "../resilience/rate-limiter.js";
-export { type HttpTransportOptions, HttpTransportServer } from "./http-transport.js";
+export { type HttpTransportOptions, HttpTransportServer } from "../transport/http.js";
 export {
   extractSingleValue,
   validateActorIdentifier,
